@@ -1,25 +1,36 @@
 {
-  "name": "cloud-disks-ui",
-  "version": "0.1.0",
+  "name": "vue-project",
+  "version": "0.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "core-js": "^3.8.3",
-    "vue": "^2.6.14",
-    "vue-router": "^3.5.1"
+    "@highlightjs/vue-plugin": "^2.1.0",
+    "@moefe/vue-aplayer": "^2.0.0-beta.5",
+    "axios": "^1.6.7",
+    "docx-preview": "^0.3.0",
+    "dplayer": "^1.27.1",
+    "element-plus": "^2.6.1",
+    "highlight.js": "^11.9.0",
+    "hls.js": "^1.5.7",
+    "js-md5": "^0.8.3",
+    "sass": "^1.71.1",
+    "sass-loader": "^14.1.1",
+    "spark-md5": "^3.0.2",
+    "vue": "^3.4.15",
+    "vue-clipboard3": "^2.0.0",
+    "vue-cookies": "^1.8.3",
+    "vue-pdf-embed": "^2.0.2",
+    "vue-router": "^4.3.0",
+    "vue3-pdfjs": "^0.1.6",
+    "xlsx": "^0.18.5"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-router": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "vue-template-compiler": "^2.6.14"
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ]
+    "@vitejs/plugin-vue": "^5.0.3",
+    "vite": "^5.0.11"
+  }
 }
