@@ -1,10 +1,38 @@
-import Vue from 'vue'
+import './assets/main.css'
+
+import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router'
+//引入Element-plus
+import ElementPlus from 'element-plus'
+import 'element-plus/dist/index.css' 
 
-Vue.config.productionTip = false
+import '@/assets/icon/iconfont.css'
+import '@/assets/base.scss'
+//引入VueCookies
+import { VueCookies } from 'vue-cookies'
 
-new Vue({
-  router,
-  render: h => h(App)
-}).$mount('#app')
+
+import Verify from './utils/Verify'
+import Message from './utils/Message'
+import Request from './utils/Request'
+import Confirm from './utils/Confirm'
+
+//自定义组件
+import Table from '@/components/Table.vue'
+
+const app = createApp(App)
+
+
+app.use(ElementPlus)
+app.use(router)
+
+app.component("Table", Table);
+
+//配置全局组件
+app.config.globalProperties.Verify=Verify;
+app.config.globalProperties.Message=Message;
+app.config.globalProperties.Request=Request;
+app.config.globalProperties.Request=Confirm;
+app.config.globalProperties.VueCookies=VueCookies;
+app.mount('#app')
